deepfake_forensics/backend/requirements.txt
fastapi
uvicorn
python-multipart
c2pa-python
torch
torchvision
numpy
scipy
opencv-python-headless
Pillow
scikit-image
matplotlib
timm
transformers
python-magic
exifread
requests
sqlalchemy
